test(t226): formalize COMMAND package fixtures

+++ tests/bx-vdm/bx_ntvdm_command_plane_v1_test.c
@@ -5,8 +5,8 @@ int main(void) {
     for(n=0;n<17;n++) { uint8_t b[4]={0xc4,0xc4,0x54,(uint8_t)n};
         bx_ntvdm_instruction_window_v1_capture(&w,b,4); bad|=!bx_ntvdm_bop_ingress_v1_classify(&w,&i);
         bad|=!bx_ntvdm_bop_provider_registry_v1_select(&i,&s); bad|=!bx_ntvdm_command_plane_v1_classify(&i,&s,&r);
-        bad|=r.service!=n || r.component==BX_NTVDM_COMMAND_COMPONENT_NONE;
-        bad|=r.disposition!=((n==3)?BX_NTVDM_COMMAND_PLANE_ORIGINAL_NOOP:((n==6||n==8||n==10)?BX_NTVDM_COMMAND_PLANE_EXPLICIT_UNAVAILABLE:BX_NTVDM_COMMAND_PLANE_DEFERRED)); }
+        bad|=r.service!=(uint8_t)n || r.component==BX_NTVDM_COMMAND_COMPONENT_NONE;
+        bad|=r.disposition!=(uint32_t)((n==3)?BX_NTVDM_COMMAND_PLANE_ORIGINAL_NOOP:((n==6||n==8||n==10)?BX_NTVDM_COMMAND_PLANE_EXPLICIT_UNAVAILABLE:BX_NTVDM_COMMAND_PLANE_DEFERRED)); }
     { uint8_t b[4]={0xc4,0xc4,0x54,17}; bx_ntvdm_instruction_window_v1_capture(&w,b,4);
       bad|=!bx_ntvdm_bop_ingress_v1_classify(&w,&i); bad|=!bx_ntvdm_bop_provider_registry_v1_select(&i,&s);
       bad|=bx_ntvdm_command_plane_v1_classify(&i,&s,&r); } return bad?1:0;
